@@ -1,5 +1,4 @@
-module.exports = {
-  devServer: {
-    disableHostCheck: true
-  }
-}
+const { defineConfig } = require('@vue/cli-service')
+module.exports = defineConfig({
+  transpileDependencies: true
+})
